[ADD] reaching the end of the control graph

+++ include/TMHead.h
@@ -8,14 +8,14 @@ class TMHead
 { // this class has to be derivided by TMTape
     public:
         TMHead() = default;
-        TMHead(unsigned int pos) : headPosition(pos) {};
+        TMHead(size_t pos) : headPosition(pos) {};
         TMHead(const TMHead&) = default;
         TMHead(TMHead&&) = default;
         virtual ~TMHead() = default;
         TMHead& operator= (const TMHead&) = default;
         TMHead& operator= (TMHead&&) = default;
-        unsigned int getHeadPosition() const { return headPosition; }
-        void setHeadPosition(unsigned int val) { headPosition = val; }
+        size_t getHeadPosition() const { return headPosition; }
+        void setHeadPosition(size_t val) { headPosition = val; }
         char getCharUnderHead() const { return *charUnderHead; }
         void setValueUnderHead(const char val) { *charUnderHead = val; }
         void setPointerForCharUnderHead(char *val) { charUnderHead = val; }
@@ -26,7 +26,7 @@ class TMHead
         virtual void doCmd (const char before, const char after, TMHeadMoveType headMove)
                 throw (MismatchCommandAndElementUnderHead, CharacterOutOfAlphabet, HeadOutOfTape) = 0;
     private:
-        unsigned int headPosition;
+        size_t headPosition;
         char *charUnderHead;
 };
